Add Barnes-Hut partition grid overlay.
The Barnes-Hut algorithm reduces the number of comparisons between
particles in an n-body simulation by recursively partitioning the
simulation space into rectangular cells so that distant groups of
particles can be treated as one particle.

I added a basic quadtree to keep track of the cells. The cell borders
are displayed as green lines. MAX_STARS_PER_CELL can be adjusted to
change the threshold at which a cell is subdivided.