Store stars in quadtree. Update cells internally.
Before this commit, the quadtree was only used as a visualization. The
update logic didn't take advantage of it. This commit utilizes the
quadtree to enable the simulation of ~2000 bodies rather than 100.

The quality of the simulation has regressed because stars are only
attracted to other stars in the same cell. Once they are also attracted
to 'virtual' stars (representing the center of mass of each other cell),
performance will go down a bit, but accuracy should go back up.

It should be noted that the quadtree root node is based around the
window size. It will need to be changed so that it always encompasses
the full simulation space, since stars outside the tree experience no
gravitational attraction.