Add a 'virtual star' for each cell.
There is now an array of special 'virtual' stars. Virtual stars can be
seen by toggling their visibility with the 'V' key. Every update, a
virtual star is created to represent the center of mass of all the stars
in each cell. The stars of each cell are then attracted to the virtual
stars of every other cell.

The whole point of this is to reduce the number of times star_attract()
needs to be called, while upholding a certain level of accuracy in the
simulation. More work needs to be done, as this currently isn't much
more efficient than the previous brute-force method. The accuracy also
leaves much to be desired.

There is also the problem that, when stars fly off the edge of the
screen, they are no longer attracted to other stars (because they aren't
in any cell of the quadtree). The root node's cell should be big enough
to enclose the two most distant stars in the simulation.

Finally, it's probably about time to clean up and organize the code so
it's easier to work with.