Add options to make testing more convenient.
`USE_TEST_SEED` seeds the PRNG with 0 every time. This results in star
positions being the same across multiple runs of the simulation.

`BRUTE_FORCE` uses the O(n^2) nested for loops to simulate gravitational
attraction. This will be useful to get an idea of the level of accuracy
achieved with the Barnes-Hut algorithm.

`RESPAWN_STARS` causes stars that leave the screen to spawn at a random
location. This will no longer be necessary once simulation field size is
independent of screen size.