Fully implement the Barnes-Hut algorithm.
See [1] for a detailed description of the Barnes-Hut Algorithm.

Each leaf node of the quad tree now holds at most one star. Internal
nodes keep track of the center of mass and total mass of their children.

The net force on each star is calculated by walking the quad tree and
attracting it to either the center of mass of an internal node, or the
star of a leaf node.

The center of mass is used only if the internal node is sufficiently far
away. The threshold that determines this is `THETA`. A `THETA` closer to
0 sacrifices performance for accuracy.

Performance hasn't improved to the degree I had hoped. Further testing
and profiling is needed to determine how best to improve efficiency.

[1]: http://arborjs.org/docs/barnes-hut