Add simulation bounding box.
The simulation no longer ends at the borders of the window. A bounding
square is calculated during every update. That box defines the
dimensions of the cell of the quad tree's root node.

There is now no need to respawn stars that go off the edge of the screen
as they are still contained within the Barnes-Hut tree.

Making the simulation independent of window and screen size is the first
step in adding view controls.